Restricted Page & User Registration(Few)

+++ base/views.py
@@ -1,6 +1,7 @@
 from django.http import HttpResponse
 from django.shortcuts import render,redirect
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from  django.contrib.auth import authenticate,login,logout
 from django.db.models import Q
@@ -15,6 +16,10 @@ from .models import Room,Topic
 # Create your views here.
 
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == "POST":
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -30,13 +35,16 @@ def loginPage(request):
         else:
             messages.error(request, 'Username of Password Does not exist!')
 
-    context = {}
+    context = {'page': page}
     return render(request,'base/login_register.html',context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
 
+def registerUser(request):
+    page = 'register'
+    return render(request,'base/login_register.html',{'page':page})
 def home(request):
     q = request.GET.get('q')if request.GET.get('q') else ''
     
@@ -52,6 +60,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()   
     if request.method == 'POST':
@@ -62,9 +71,16 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html',context)
 
+
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user == room.host:
+        return HttpResponse('You are not allowed to update this room')
+
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -74,8 +90,14 @@ def updateRoom(request,pk):
     return render(request,'base/room_form.html',context)
 
 
+
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user == room.host:
+        return HttpResponse('You are not allowed to delete this room')
+
     if request.method=="POST":
         room.delete()
         return redirect('home')
